:art: refactoring WIP
- 폴더 구조 개선 진행중
- DTO 폴더만 개선하면 일차적인 개선 진행
- response가 사용되지 않은 부분 수정 예정
