+++ src/course/index.ts
@@ -1,0 +1,3 @@
+export * from './repository';
+export * from './service';
+export * from './courseRouter';
