+++ src/course/repository/index.ts
@@ -1,0 +1,5 @@
+export * from './courseLogRepository';
+export * from './courseRepository';
+export * from './quizLogRepository';
+export * from './quizRepository';
+export * from './statisticsRepository';
