+++ src/dto/request/index.ts
@@ -7,3 +7,4 @@ export * from './SubmitQuizRequest';
 export * from './UpdatePasswordRequest';
 export * from './GetPostQuery';
 export * from './UpdateStatisticsRequest';
+export * from './UpdateUserRequest';
