+++ src/middlewares/index.ts
@@ -1,3 +1,4 @@
 export * from './authMiddleware';
 export * from './sessionMiddleware';
+export * from './uploaderMiddleware';
 export * from './validationMiddleware';
