+++ src/post/index.ts
@@ -1,0 +1,3 @@
+export * from './repository';
+export * from './postRouter';
+export * from './postService';
