+++ src/post/repository/index.ts
@@ -1,0 +1,2 @@
+export * from './commentRepository';
+export * from './postRepository';
