--- src/repositories/index.ts
@@ -1,8 +1,0 @@
-export * from './userRepository';
-export * from './commentRepository';
-export * from './courseRepository';
-export * from './courseLogRepository';
-export * from './postRepository';
-export * from './quizRepository';
-export * from './quizLogRepository';
-export * from './statisticsRepository';
