--- src/services/index.ts
@@ -1,6 +1,0 @@
-export * from './courseService';
-export * from './courseLogService';
-export * from './postService';
-export * from './userService';
-export * from './quizService';
-export * from './quizLogService';
